+++ new/coding/src/database/migrations/2025_06_19_120035_add_kota_to_programs_table.php
@@ -6,23 +6,21 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
         Schema::table('programs', function (Blueprint $table) {
-            $table->string('kota')->after('kategori');
+            if (!Schema::hasColumn('programs', 'kota')) {
+                $table->string('kota')->after('kategori');
+            }
         });
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
         Schema::table('programs', function (Blueprint $table) {
-            $table->dropColumn('kota');
+            if (Schema::hasColumn('programs', 'kota')) {
+                $table->dropColumn('kota');
+            }
         });
     }
 };
